pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "effectful"
version = "0.2.1"
description = "Metaprogramming infrastructure"
readme = "README.rst"
license = "Apache-2.0"
requires-python = ">=3.12"
authors = [
    { name = "Basis" },
]
keywords = [
    "machine learning",
    "statistics",
    "probabilistic programming",
    "bayesian modeling",
    "pytorch",
]
classifiers = [
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = []

[project.urls]
Homepage = "https://www.basis.ai/"
Source = "https://github.com/BasisResearch/effectful"
"Bug Tracker" = "https://github.com/BasisResearch/effectful/issues"

[project.optional-dependencies]
torch = ["torch", "dm-tree"]
pyro = ["pyro-ppl>=1.9.1", "dm-tree"]
jax = ["jax", "dm-tree"]
numpyro = ["numpyro>=0.19", "dm-tree"]
openai = ["openai", "pillow"]
anthropic = ["anthropic", "pillow"]
docs = [
    "effectful[torch,pyro,jax,numpyro]",
    "sphinx",
    "sphinxcontrib-bibtex",
    "sphinx_rtd_theme",
    "myst-parser",
    "nbsphinx",
    "sphinx_autodoc_typehints",
    "pypandoc_binary",
]
test = [
    "effectful[torch,pyro,jax,numpyro,docs]",
    "pytest",
    "pytest-cov",
    "pytest-xdist",
    "pytest-benchmark",
    "mypy",
    "ruff",
    "nbval",
    "nbqa",
]

[dependency-groups]
dev = ["effectful[torch,pyro,jax,numpyro,openai,anthropic,docs,test]"]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["F", "I", "PERF", "UP"]

[tool.pytest.ini_options]
addopts = "--doctest-modules"
testpaths = ["./effectful", "./tests"]

[tool.mypy]
ignore_missing_imports = true
warn_unused_ignores = true
exclude = "build"
